[MOVE-ONLY] Move CBudgetDB to budgetdb.h/cpp
