+++ src/init.cpp
@@ -17,6 +17,7 @@
 #include "activemasternode.h"
 #include "addrman.h"
 #include "amount.h"
+#include "budget/budgetdb.h"
 #include "checkpoints.h"
 #include "compat/sanity.h"
 #include "consensus/upgrades.h"
@@ -26,7 +27,6 @@
 #include "httprpc.h"
 #include "invalid.h"
 #include "key.h"
-#include "masternode-budget.h"
 #include "masternode-payments.h"
 #include "masternodeconfig.h"
 #include "masternodeman.h"
